data layer restructured using a repository interface
