--- Data/Core/Abstracts/DataServiceAdapterAbstract.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Data\Core\Abstracts;
-
-abstract class DataServiceAdapterAbstract
-{
-    // Adapter abstract for data service apis.
-}
