+++ Data/Core/Adapters/DB/MySQL/MySQLAdapter.php
@@ -125,8 +125,10 @@ class MySQLAdapter extends DBAdapterAbstract
      * COUNT from a table.
      *
      * @param $strTable
-     * @param array $arrWhere
-     * @param array $arrJoins
+     * @param array $arrWhere [['column_1', '=', 'value'],['column_2', '=', 'value', 'OR'],['column_2', 'LIKE', '%value%'],
+     *                         ['column_2', 'IN', [1, 2, 3]],['column_2', 'BETWEEN', [value_1, value_2]]]
+     * @param array $arrJoins [['LEFT JOIN', 'table_1', 'table_1.column', 'other_table.column'],
+     *                         ['JOIN', 'table_2', 'table_2.column', 'other_table.column']]
      * @return int
      */
     public function count($strTable, array $arrWhere = [], array $arrJoins = [])
