--- Data/Core/Adapters/Service/Example/ExampleDataServiceAdapter.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Data\Core\Adapters\Service\Example;
-
-use Data\Core\Abstracts\DataServiceAdapterAbstract;
-
-class ExampleDataServiceAdapter
-{
-    // Adapter abstract for data service apis.
-}
