--- Data/Core/Exceptions/Types/DataServiceException.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Data\Core\Exceptions\Types;
-
-class DataServiceException extends \Exception
-{
-
-}
