+++ Data/Repositories/HelloRepository.php
@@ -3,182 +3,163 @@
 namespace Data\Repositories;
 
 use Data\Core\Abstracts\RepositoryAbstract;
+use Data\Core\Interfaces\RepositoryInterface;
 use Data\Models\Hello;
 use Data\Models\User;
 
-class HelloRepository extends RepositoryAbstract
+class HelloRepository extends RepositoryAbstract implements RepositoryInterface
 {
-    public function selectHello($arrFilters)
-	{
-	    // define filter mappings to table columns ___
-//        $arrMappings = [
-//            'trip_status' => null,
-//            'passenger_phone' => 'phone',
-//            'type' => null, // null is mapped to fields that are static filters
-//        ];
+
+    /**
+     * Select that can be manipulated using a filters array.
+     *
+     * @param array $arrFilters [[ <string>'field', <integer>1 (operator), <mixed>'value' ]
+     *                          [ <string>'field', <integer>1 (operator), <mixed>'value' ]]
+     * @return array
+     */
+    public function select($arrFilters = [])
+    {
+        // define filter mappings to table columns ___
+        //      (NOTE: 'filter_name' => 'mapped_table_column_name',
+        //              'another_filter_name' => null
+        //              null is mapped to fields that will be treated as static filters)
+        $arrMappings = [
+
+        ];
 
         // rebuild the filters array to be compatible with the repository ___
-//        $arrFilters = $this->rebuildFilter($arrFilters, $arrMappings);
+        $arrFilters = $this->rebuildFilter($arrFilters, $arrMappings);
 
+        // where conditions that should be assigned regardless of the presence or the absence of filters ___
+        $arrWhere = [
 
-	    $arrWhere = [
-//	          ['id', '!=', 1],
-//              ['status', 'IN', [1, 2, 3, 4, 5], false],
-//              [[['col1', '!=', 2], ['col2', '!=', 3]], true],
-//              [[['col3', 'IN', [1,2,3]], ['col126548', '!=', 3]]]
-           ];
+        ];
 
-	    // add static where conditions ___
+        // add where conditions based on static filters ___
 //        if($arrFilters['static']['trip_status'] == 1)
 //        {
 //            $arrWhere[] = ['status', '=', 10, true];
 //        }
 
-	    // add dynamic where conditions ___
-//        $arrWhere = array_merge($arrWhere, $arrFilters['dynamic']);
+        // combine where conditions generated by dynamic conditions sent through filters to the main set of where conditions ___
+        $arrWhere = array_merge($arrWhere, $arrFilters['dynamic']);
 
-
+        // add joins ___
         $arrJoins = [
-//            ['LEFT JOIN', 'table_1 T1', 'T1.column', 'T.column'],
-//            ['JOIN', 'table_2 T2', 'T2.column', 'T.column'],
+//            ['LEFT JOIN', User::TABLE . ' U', 'U.id', 'H.user_id'],
+//            ['LEFT JOIN', Type::TABLE . ' T', 'T.id', 'H.type_id'],
         ];
 
+        // add ordering ___
         $arrOrder = [
-//            ['id', 'ASC'],
-//            ['status', 'DESC'],
+//            ['column_1' => 'ASC'],
+//            ['column_2', 'DESC']
         ];
 
+        // add fetch limit ___
         $arrLimit = [
-//            0, //offset
-//            10, // limit
+//            offset,
+//            limit
         ];
 
+        // add return columns ___
         $arrColumns = [
-//            'id AS identifier',
-//            'col_1 AS column_1',
+//            'H.id',
+//            'H.name',
+//            'H.description',
         ];
 
-        return $this->db->select(User::TABLE, $arrJoins, $arrWhere, $arrOrder, $arrLimit, $arrColumns);
-	}
+        return $this->db->select(Hello::TABLE . ' AS H', $arrJoins, $arrWhere, $arrOrder, $arrLimit, $arrColumns);
+    }
 
 
-	public function countHello($arrFilter)
+    /**
+     * Select a single item using its primary key value.
+     *
+     * @param $intId
+     * @return array
+     */
+    public function selectItem($intId)
     {
+        // should map item identifier to primary key field ___
         $arrWhere = [
-//	          ['id', '!=', 1],
-//	          ['col_1', 'LIKE', 'sri lanka%'],
-//            ['date', 'BETWEEN', ['2017-01-01', '2017-01-02']],
-//            ['status', 'IN', [1, 2, 3, 4, 5]],
+            ['H.id', '=', $intId]
         ];
 
+        // add joins ___
+        $arrJoins = [
+//            ['LEFT JOIN', User::TABLE . ' U', 'U.id', 'H.user_id'],
+//            ['LEFT JOIN', Type::TABLE . ' T', 'T.id', 'H.type_id'],
+        ];
+
+        // add return columns ___
+        $arrColumns = [
+//            'H.id',
+//            'H.name',
+//            'H.description',
+        ];
+
+        return $this->db->select(Hello::TABLE . ' AS H', $arrJoins, $arrWhere, [], [], $arrColumns);
+    }
+
+
+    /**
+     * Select data from the base table by different criteria.
+     *
+     * @param array $arrWhere [['column_1', '=', 'value'],['column_2', '=', 'value', 'OR'],['column_2', 'LIKE', '%value%'],
+     *                         ['column_2', 'IN', [1, 2, 3]],['column_2', 'BETWEEN', [value_1, value_2]]]
+     * @param array $arrOrder [['column_1' => 'ASC'], ['column_2', 'DESC']]
+     * @param array $arrLimit [offset, limit]
+     * @param array $arrColumns ['column_1', 'column_2', ...]
+     * @return array
+     */
+    public function selectBy($arrWhere = [], $arrOrder = [], $arrLimit = [], $arrColumns = [])
+    {
+        return $this->db->select(Hello::TABLE, [], $arrWhere, $arrOrder, $arrLimit, $arrColumns);
+    }
+
+
+    /**
+     * Get a count from the base table by different criteria.
+     *
+     * @param array $arrWhere [['column_1', '=', 'value'],['column_2', '=', 'value', 'OR'],['column_2', 'LIKE', '%value%'],
+     *                         ['column_2', 'IN', [1, 2, 3]],['column_2', 'BETWEEN', [value_1, value_2]]]
+     * @return integer
+     */
+    public function count($arrWhere = [])
+    {
         return $this->db->count(Hello::TABLE, $arrWhere);
     }
 
 
-    public function insertHello()
+    /**
+     * Insert a record in to the base table.
+     *  (NOTE: The db adapter also supports bulk insets through
+     *         insertBulk($strTable, $arrColumns, $arrValues) method)
+     *
+     * @param $arrRecord ['column_1' = > value_1, 'column_2' => value_2, ...]
+     * @return array
+     */
+    public function insert($arrRecord)
     {
-        $arrRecord = [
-            'col_1' => 'value_1',
-            'col_2' => 'value_2',
-            'col_3' => 'value_3',
-        ];
-
         return $this->db->insert(Hello::TABLE, $arrRecord);
     }
 
 
-    public function insertMultipleHello()
+    /**
+     * Update a record in the base table.
+     *
+     * @param $intId
+     * @param $arrSet ['column_1' => value_1, 'column_2' => value_2, ...]
+     * @return array
+     */
+    public function update($intId, $arrSet)
     {
-        $arrColumns = [
-            'col_1',
-            'col_2',
-            'col_3',
-        ];
-
-        $arrValues = [
-            ['val_11', 'val_12', null],
-            ['val_21', 'val_22', null],
-            ['val_21', 'val_22', null],
-        ];
-
-        $this->db->transBegin();
-
-        try
-        {
-            $result = $this->db->insertBulk(Hello::TABLE, $arrColumns, $arrValues);
-
-            $this->db->transCommit();
-
-            return $result;
-        }
-        catch(\Exception $e)
-        {
-            $this->db->transRollback();
-
-            throw $e;
-        }
-    }
-
-
-    public function updateHello()
-    {
-        $arrSet = [
-            'col_1' => 'updated_again2',
-            'col_2' => 'updated_again2',
-            'col_3' => 'updated_again2',
-        ];
-
+        // should map item identifier to primary key field ___
         $arrWhere = [
-        ['id', 'IN', [1, 2]],
-//	        ['col_1', 'LIKE', 'sri lanka%'],
-//            ['date', 'BETWEEN', ['2017-01-01', '2017-01-02']],
-//            ['status', 'IN', [1, 2, 3, 4, 5]],
+            ['id', '=', $intId]
         ];
 
-        $this->db->transBegin();
-
-        try
-        {
-            $result = $this->db->update(Hello::TABLE, $arrSet, $arrWhere);
-
-            $this->db->transCommit();
-
-            return $result;
-        }
-        catch(\Exception $e)
-        {
-            $this->db->transRollback();
-
-            throw $e;
-        }
+        return $this->db->update(Hello::TABLE, $arrSet, $arrWhere);
     }
-
-
-    public function deleteHello()
-    {
-        $arrWhere = [
-        ['id', '=', 5],
-//	        ['col_1', 'LIKE', 'sri lanka%'],
-//            ['date', 'BETWEEN', ['2017-01-01', '2017-01-02']],
-//            ['status', 'IN', [1, 2, 3, 4, 5]],
-        ];
-
-        $this->db->transBegin();
-
-        try
-        {
-            $result = $this->db->delete(Hello::TABLE, $arrWhere);
-
-            $this->db->transCommit();
-
-            return $result;
-        }
-        catch(\Exception $e)
-        {
-            $this->db->transRollback();
-
-            throw $e;
-        }
-    }
-
 }
