--- Domain/Core/Exceptions/Types/LogicServiceException.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Domain\Core\Exceptions\Types;
-
-class LogicServiceException extends \Exception
-{
-
-}
