+++ api/Core/Adapters/Service/GuzzleAdapter.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Api\Core\Adapters\Service;
+
+class GuzzleAdapter
+{
+
+}
